# Trying out Span<char> and Memory<char> for improved performance