--- a/MarkdownLinkCheckLogParser/src/MarkdownLinkCheckLogParserCli/GitHub/GitHubLogLine.cs
+++ b/MarkdownLinkCheckLogParser/src/MarkdownLinkCheckLogParserCli/GitHub/GitHubLogLine.cs
@@ -1,32 +1,23 @@
 namespace MarkdownLinkCheckLogParserCli.GitHub;
 
-internal class GitHubLogLine
+internal readonly struct GitHubLogLine
 {
-    private readonly string _logLine;
-    private readonly Lazy<ReadOnlyMemory<char>> _logLineWithoutTimestamp;
-
-    public GitHubLogLine(string logLine)
+    public GitHubLogLine(ReadOnlyMemory<char> line)
     {
-        _logLine = logLine;
-        _logLineWithoutTimestamp = new Lazy<ReadOnlyMemory<char>>(() => TrimTimestamp(logLine));
+        Value = RemoveTimestamp(line);
     }
 
-    public static implicit operator string(GitHubLogLine logLine)
+    public ReadOnlyMemory<char> Value { get; }
+
+    public bool IsEmpty => Value.IsEmpty;
+
+    public static implicit operator ReadOnlyMemory<char>(GitHubLogLine line) => line.Value;
+
+    private static ReadOnlyMemory<char> RemoveTimestamp(ReadOnlyMemory<char> line)
     {
-        return logLine._logLine;
-    }
-
-    public ReadOnlyMemory<char> WithoutTimestamp => _logLineWithoutTimestamp.Value;
-
-    private static ReadOnlyMemory<char> TrimTimestamp(string logLine)
-    {
-        var indexOfFirstSpace = logLine.IndexOf(' ', StringComparison.InvariantCulture);
-        var spanWithoutTimestamp = logLine.AsMemory(start: indexOfFirstSpace);
-        
-
-        //var splits = logLine.Split(" ");
-        //return string.Join(" ", splits.Skip(1));
-
-        return spanWithoutTimestamp;
+        var indexOfSpace = line.Span.IndexOf(' ');
+        return line
+            .Slice(indexOfSpace + 1)
+            .TrimStart();
     }
 }