--- a/MarkdownLinkCheckLogParser/src/MarkdownLinkCheckLogParserCli/GitHub/GitHubStepLog.cs
+++ b/MarkdownLinkCheckLogParser/src/MarkdownLinkCheckLogParserCli/GitHub/GitHubStepLog.cs
@@ -1,0 +1,19 @@
+namespace MarkdownLinkCheckLogParserCli.GitHub;
+
+internal class GitHubStepLog
+{
+    private readonly Memory<char> _log;
+
+    public GitHubStepLog(Memory<char> log)
+    {
+        _log = log;
+    }
+
+    public IEnumerable<GitHubLogLine> GetLogLines()
+    {
+        foreach (ReadOnlyMemory<char> line in _log.SplitLines())
+        {
+            yield return new GitHubLogLine(line);
+        }
+    }
+}