--- a/MarkdownLinkCheckLogParser/src/MarkdownLinkCheckLogParserCli/GitHub/GitHubWorkflowRunLogs.cs
+++ b/MarkdownLinkCheckLogParser/src/MarkdownLinkCheckLogParserCli/GitHub/GitHubWorkflowRunLogs.cs
@@ -9,9 +9,9 @@ internal class GitHubWorkflowRunLogs
         _gitHubHttpClient = gitHubHttpClient.NotNull();
     }
 
-    public async Task<ReadOnlyMemory<char>> GetWorkflowRunLogForStepAsync(string repo, string runId, string jobName, string stepName)
+    public async Task<ZipArchiveEntry> GetWorkflowRunLogForStepAsync(string repo, string runId, string jobName, string stepName)
     {
-        using var workflowRunLogsZip = await _gitHubHttpClient.DownloadWorkflowRunLogsAsync(repo, runId);
+        var workflowRunLogsZip = await _gitHubHttpClient.DownloadWorkflowRunLogsAsync(repo, runId);
         var markdownLinkCheckLogsZipEntrys = workflowRunLogsZip.Entries.
             Where(e => e.FullName.Contains($"{jobName}/", StringComparison.Ordinal) && e.Name.Contains(stepName, StringComparison.Ordinal))
             .ToList();
@@ -27,11 +27,6 @@ internal class GitHubWorkflowRunLogs
             throw new Exception("TODO more than one match");
         }
 
-        var stepLogZipEntry = markdownLinkCheckLogsZipEntrys[0];
-        using var stepLogStream = stepLogZipEntry.Open();
-        using var streamReader = new StreamReader(stepLogStream, Encoding.UTF8);
-        var stepLogMemory = new Memory<char>(new char[stepLogZipEntry.Length]);
-        await streamReader.ReadAsync(stepLogMemory);
-        return stepLogMemory;
+        return markdownLinkCheckLogsZipEntrys[0];      
     }
 }