--- a/MarkdownLinkCheckLogParser/src/MarkdownLinkCheckLogParserCli/GlobalUsings.cs
+++ b/MarkdownLinkCheckLogParser/src/MarkdownLinkCheckLogParserCli/GlobalUsings.cs
@@ -1,4 +1,5 @@
 global using System.Diagnostics.CodeAnalysis;
+global using System.Globalization;
 global using System.IO.Compression;
 global using System.Net.Http.Headers;
 global using System.Runtime.CompilerServices;