--- a/MarkdownLinkCheckLogParser/src/MarkdownLinkCheckLogParserCli/GlobalUsings.cs
+++ b/MarkdownLinkCheckLogParser/src/MarkdownLinkCheckLogParserCli/GlobalUsings.cs
@@ -1,10 +1,8 @@
 global using System.Diagnostics.CodeAnalysis;
-global using System.Globalization;
 global using System.IO.Compression;
 global using System.Net.Http.Headers;
 global using System.Runtime.CompilerServices;
 global using System.Text;
-global using System.Text.RegularExpressions;
 global using CliFx;
 global using CliFx.Attributes;
 global using CliFx.Exceptions;
@@ -12,4 +10,5 @@ global using CliFx.Extensibility;
 global using CliFx.Infrastructure;
 global using MarkdownLinkCheckLogParserCli.ArgumentValidations;
 global using MarkdownLinkCheckLogParserCli.GitHub;
+global using MarkdownLinkCheckLogParserCli.MarkdownLinkCheck;
 global using MarkdownLinkCheckLogParserCli.MarkdownLinkCheck.LogLines;