--- a/MarkdownLinkCheckLogParser/src/MarkdownLinkCheckLogParserCli/MarkdownLinkCheck/LogLines/IMarkdownLinkCheckLogLine.cs
+++ b/MarkdownLinkCheckLogParser/src/MarkdownLinkCheckLogParserCli/MarkdownLinkCheck/LogLines/IMarkdownLinkCheckLogLine.cs
@@ -1,0 +1,6 @@
+namespace MarkdownLinkCheckLogParserCli.MarkdownLinkCheck.LogLines;
+
+internal interface IMarkdownLinkCheckLogLine
+{
+    void Handle(ParserState state);
+}