--- a/MarkdownLinkCheckLogParser/src/MarkdownLinkCheckLogParserCli/MarkdownLinkCheck/LogLines/MarkdownLinkCheckLogLine.cs
+++ b/MarkdownLinkCheckLogParser/src/MarkdownLinkCheckLogParserCli/MarkdownLinkCheck/LogLines/MarkdownLinkCheckLogLine.cs
@@ -1,8 +1,0 @@
-namespace MarkdownLinkCheckLogParserCli.MarkdownLinkCheck.LogLines;
-
-internal abstract class MarkdownLinkCheckLogLine
-{
-    public static NotValidLogLine NotValidLogLine { get; } = new NotValidLogLine();
-
-    public abstract void Handle(ParserState state);
-}