--- a/MarkdownLinkCheckLogParser/tests/MarkdownLinkCheckLogParserCli.Benchmarks/ParsingBenchmark.cs
+++ b/MarkdownLinkCheckLogParser/tests/MarkdownLinkCheckLogParserCli.Benchmarks/ParsingBenchmark.cs
@@ -24,7 +24,7 @@ public class ParsingBenchmark
         };
         var console = new FakeInMemoryConsole();
         await command.ExecuteAsync(console);
-    } 
+    }
 }
 
 public class InMemoryGitHubWorkflowRunHandler : DelegatingHandler