… MarkdownLinkCheckLogParser/src/MarkdownLinkCheckLogParserCli/MarkdownLinkCheck/MemoryExtensions.cs
… MarkdownLinkCheckLogParser/src/MarkdownLinkCheckLogParserCli/MarkdownLinkCheck/ParserState.cs
… MarkdownLinkCheckLogParser/tests/MarkdownLinkCheckLogParserCli.Benchmarks/Program.cs