This PR changes the parsing on the GitHub step log from using `string` and `Regex` to using `Memory<char>` and `Span<char>`.

## Benchmarks

The version before this PR has the following results on the Benchmark DotNet project:

| Method |     Mean |    Error |   StdDev |   Gen 0 |   Gen 1 | Allocated |
|------- |---------:|---------:|---------:|--------:|--------:|----------:|
|  Parse | 940.7 μs | 16.78 μs | 14.88 μs | 95.7031 | 23.4375 |    592 KB |

The version on commit 2fbbbf4409af2d162abd1ef94353e8cc41c2e961 only uses Span<char> and had the following results:

| Method |     Mean |   Error |  StdDev |   Gen 0 |   Gen 1 | Allocated |
|------- |---------:|--------:|--------:|--------:|--------:|----------:|
|  Parse | 190.4 μs | 1.84 μs | 1.63 μs | 23.6816 | 11.7188 |    147 KB |

The final version of this PR uses Memory<char> instead of just Span<char> and had the following results:

| Method |     Mean |   Error |  StdDev |   Gen 0 |   Gen 1 | Allocated |
|------- |---------:|--------:|--------:|--------:|--------:|----------:|
|  Parse | 289.1 μs | 5.54 μs | 5.18 μs | 25.3906 | 12.6953 |    157 KB |

## Why did I go with the version with `Memory<char>` if it was slower and allocated more memory than the fully `Span<char>` version ?

Using `Span` comes with restrictions, you can't use it on `async` methods, you can't use it as a field on a class, you can't return it on iterators, you can't use it on structs unless they are `ref struct`, etc.

You can see that the version of the code on commit 2fbbbf4409af2d162abd1ef94353e8cc41c2e961 is very rough in terms of conveying meaning of what the code is doing. Although this version could surely be improved I preferred for now to use `Memory<char>` and keep the existing classes and code structure that I believe better conveys what the code is doing and is more maintainable in the long run I believe.

Don't get me wrong, using the version with only `Span<char>` and polishing the code a bit would make it more than understandable and  maintainable for a project of this size. In the same way that the difference in performance from the version with `Span<char>` to the version with `Memory<char>` doesn't really matter for this project.

So in the end I chose what I believe is a better code structure to the detriment of a bit of performance.

## If performance is not that important in this project why did you do this ?

I did it as a learning experience. I got a real life project where I could play with `Span` and `Memory` so why not ?

## PS

This PR also includes small improvements to the code structure.